DO NOT MERGE Ensure the associated call count is accurately tracked for RCS.
This CL ensures that the associatedCallCount of both Connection Services and Remote Connection Services is accurately being tracked. Specifically this CL addresses the following 2 scenarios: (1) when the user merges 2 Fi calls into a conference and (2) when the user disconnects an active TMO call while there is a Fi call on hold. This is primarily accomplished by ensuring that both the primary CS associatedCallCount and the RCS associatedCallCount are incremented when a Fi (or other call dependent on an RCS) call is created.

Bug: 286154316
Test: manual
Change-Id: I05104e89a7620165a6c00e2d276509c055b6324d
(cherry picked from commit 45c65d16d0b178744372a62134fe5ba0c4487cdd)