Clear communication device after audio ops finish
Modifies the existing logic to ensure that after the call ends, the
currently requested device for communication is cleared once the audio
operations are complete.

A bug had surfaced for an uncommon case where the audio mode would enter
the unfocused state before a bt device disconnected, resulting in
Telecom clearing the bt device from communication. However, once the bt
device was disconnected, the route entered the ActiveEarpieceRoute,
which caused the earpiece to be set for communication (but it never got
cleared). This CL moves the existing logic to after when
AUDIO_OPERATIONS_COMPLETE is processed in the UnfocusedState instead to
resolve the timing issues with the BT device disconnect and the audio
mode entering the unfocused state.

Bug: 315865533
Test: Manual
Change-Id: If88dbedb9ef505b1508273e8ad98a463b791ed71