Resolve Ringer#getUri Telecom usages
As part of mainline prep, resolve the Telecom references to
Ringtone#getUri (hidden). This is primarily used for logging so we can
remove these references. In the case of finding the vibration effect
associated with the Uri, this can be referenced from the uri determined
in RingtoneFactory#getHapticOnlyRingtone or RingtoneFactory#getRingtone.

Bug: 307815630
Bug: 311773409
Test: atest TelecomUnitTests
Change-Id: I1e6019156513f0821aa162efc412acff20962ab6