Cache callbacks that can be ready before the service is set
After digging through some failed flaky call audio tests, I saw that the
CallEndpointController was not calling the callbacks because the
TransactionalServiceWrapper was null at the time the callbacks were
ready to be called.  This is a timing issue that is hard to replicate.

To mitigate, the ServiceWrappers should cache the callbacks and once the
service is set, the callbacks should be sent.

Fixes: 321369729
Test: atest com.android.server.telecom.tests.CallTest
Change-Id: I923b5ddc9e3f801113c56f8b710b597f6d2efd0b