Correct erroneous call logging failure reporting.
There was some issue with the call log "verification" code we added wrt
to multiuser functionality.  This code was causing a lot of false reports
of call logging issues.  In practice this code was added as a speculative
way to diagnose call logging issues on the assumption the provider was
not actually logging.  As it turned out the logging issues we had in
the past were not related at all.  Checking for a valid URI after logging
is enough to know that a call was logged.  Internal logging in the
provider itself would point out to DB errors and things like that.

Test: Manually reproduce logging issue and confirm it is fixed.
Fixes: 330282747
Change-Id: I2a72dcbf2e4b4f9a520cfdb47ef5aca5a6bca01c