Create toast on main thread executor.
With the recent mainline changes to resolve the hidden reference to
Toast#makeText, the new method tries to use Looper#myLooper in creating
the toast. However, this ends up throwing an NPE because we are running
this off of a binder thread. Instead, we need to explicitly run this
command in an exeuctor running on the main thread.

Bug: 330526426
Test: Manual (tested that the new method isn't causing crashes and the
toasts are successfully created)

Change-Id: Iea1223e9c5cdaef7afd456b2605474eb00aaf623