Move carrier in-service check to after E911 timeout.
Instead of choosing whether to add a timeout when iterating through the phone accounts to see if we should make the CallAttempt, we now assign the timeout and check for the Voice capability at the end. This way, the connection service can keep trying if the CallManager is unable to handle the call after the 25 seconds.

Bug: 333816574
Test: atest CreateConnectionProcessorTest
Test: manually deployed and tested in a shieldbox with Fi's connection service to prevent radio connectivity
Change-Id: I8ffca4cb948d18f37a6b93f6eec2f82630c7ddf6