Serial & Parallel transactions should not override the failure values
Every failed transaction was returning CallException.ERROR_CODE_UNKNOWN due to the Serial and Parallel transactions overriding the
VoipCallTransactionResult values. To better inform the client of why a
particular transaction failed, the code and message should be sent to
the client.

Additionally, the Telecom codebase should ONLY use fail
codes defined in the CallException class when failing a transaction. The old VoipCallTransactionResult.RESULT_FAILED value has been removed in favor of using a defined value in the CallException class.

Flag: NONE <fixes a regression>
Fixes: 313461479
Test: CTS
Change-Id: I3c1b34e40890f74fa91c7475d74b21f5ca9b96aa