Fix issue with unmute when BT moves to Quiescent.
The transitRouteBeforeAudioDisconnectBt flagged path does not use the
reinitialize method and instead moves directly to the quiescent BT route.
It does this because the reinitialize method will always revert back to
earpiece route.  BT is a "sticky" route, so we just move back to quiescent
BT if it is still active.

However, because this flagged path doesn't use reinitialize, we never turn
off mute during the state transition.  The resetMuteWhenEnteringQuiescentBtRoute
flag attempted to fix that but put it in the ENTER for the
QuiescentBluetooth route.  The downside to that is that setMuteOn method
does notthing when the current state is a quiescent state.

So, we'll move that logic before the transition to quiescent.


Bug: 336718198
Test: Manual test.
Test: Expanded testQuiescentBluetoothRouteResetMute to verify mute happens.
Change-Id: If97ecc8e2440836ae971be0a1eedae8230d8df3b